Define some models inside the sandbox and some outside
